+++ src/main/java/com/box/sdk/BoxLegalHoldPolicy.java
@@ -72,7 +72,7 @@ public class BoxLegalHoldPolicy extends BoxResource {
      * @return          information about the Legal Hold Policy created.
      */
     public static BoxLegalHoldPolicy.Info create(BoxAPIConnection api, String name) {
-        return create(api, name, null, null, null);
+        return createOngoing(api, name, null);
     }
 
     /**
@@ -98,6 +98,29 @@ public class BoxLegalHoldPolicy extends BoxResource {
         }
         if (filterEndedAt != null) {
             requestJSON.add("filter_ended_at", BoxDateFormat.format(filterEndedAt));
+        }
+        request.setBody(requestJSON.toString());
+        BoxJSONResponse response = (BoxJSONResponse) request.send();
+        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        BoxLegalHoldPolicy createdPolicy = new BoxLegalHoldPolicy(api, responseJSON.get("id").asString());
+        return createdPolicy.new Info(responseJSON);
+    }
+
+    /**
+     * Creates a new ongoing Legal Hold Policy.
+     * @param api               the API connection to be used by the resource.
+     * @param name              the name of Legal Hold Policy.
+     * @param description       the description of Legal Hold Policy.
+     * @return                  information about the Legal Hold Policy created.
+     */
+    public static BoxLegalHoldPolicy.Info createOngoing(BoxAPIConnection api, String name, String description) {
+        URL url = ALL_LEGAL_HOLD_URL_TEMPLATE.build(api.getBaseURL());
+        BoxJSONRequest request = new BoxJSONRequest(api, url, "POST");
+        JsonObject requestJSON = new JsonObject()
+                .add("policy_name", name)
+                .add("is_ongoing", true);
+        if (description != null) {
+            requestJSON.add("description", description);
         }
         request.setBody(requestJSON.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
@@ -331,6 +354,11 @@ public class BoxLegalHoldPolicy extends BoxResource {
         private String releaseNotes;
 
         /**
+         * @see #getIsOngoing()
+         */
+        private Boolean isOngoing;
+
+        /**
          * Constructs an empty Info object.
          */
         public Info() {
@@ -461,6 +489,13 @@ public class BoxLegalHoldPolicy extends BoxResource {
         }
 
         /**
+         * @return boolean indicating whether the policy will continue applying to files based on events, indefinitely.
+         */
+        public Boolean getIsOngoing() {
+            return this.isOngoing;
+        }
+
+        /**
          * {@inheritDoc}
          */
         @Override
@@ -502,6 +537,8 @@ public class BoxLegalHoldPolicy extends BoxResource {
                     this.filterStartedAt = BoxDateFormat.parse(value.asString());
                 } else if (memberName.equals("filter_ended_at")) {
                     this.filterEndedAt = BoxDateFormat.parse(value.asString());
+                } else if (memberName.equals("is_ongoing")) {
+                    this.isOngoing = value.asBoolean();
                 }
             } catch (ParseException e) {
                 assert false : "A ParseException indicates a bug in the SDK.";