+++ src/main/java/com/box/sdk/FileUploadParams.java
@@ -8,6 +8,7 @@ import java.util.Date;
  */
 public class FileUploadParams {
     private InputStream content;
+    private UploadFileCallback uploadFileCallback;
     private String name;
     private Date created;
     private Date modified;
@@ -35,6 +36,24 @@ public class FileUploadParams {
      */
     public FileUploadParams setContent(InputStream content) {
         this.content = content;
+        return this;
+    }
+
+    /**
+     * @return content writer callback.
+     */
+    public UploadFileCallback getUploadFileCallback() {
+        return this.uploadFileCallback;
+    }
+
+    /**
+     * Sets the content writer callback.
+     *
+     * @param uploadFileCallback callback called when file upload starts.
+     * @return         this FileUploadParams object for chaining.
+     */
+    public FileUploadParams setUploadFileCallback(UploadFileCallback uploadFileCallback) {
+        this.uploadFileCallback = uploadFileCallback;
         return this;
     }