+++ src/test/java/com/box/sdk/BoxAPIConnectionTest.java
@@ -662,4 +662,37 @@ public class BoxAPIConnectionTest {
 
         BoxFile.Info info = new BoxFile(api, "98765").getInfo();
     }
+
+    @Test
+    @Category(UnitTest.class)
+    public void shouldUseGlobalMaxRequests() {
+
+        int defaultMaxRequests = BoxGlobalSettings.getMaxRequestAttempts();
+        int newMaxRequests = defaultMaxRequests + 5;
+        BoxGlobalSettings.setMaxRequestAttempts(newMaxRequests);
+
+        BoxAPIConnection api = new BoxAPIConnection("");
+        assertEquals(newMaxRequests, api.getMaxRequestAttempts());
+
+        // Set back the original number to not interfere with other test cases
+        BoxGlobalSettings.setMaxRequestAttempts(defaultMaxRequests);
+    }
+
+    @Test
+    @Category(UnitTest.class)
+    public void shouldUseInstanceTimeoutSettings() throws MalformedURLException {
+
+        int instanceConnectTimeout = BoxGlobalSettings.getConnectTimeout() + 1000;
+        int instanceReadTimeout = BoxGlobalSettings.getReadTimeout() + 1000;
+
+        BoxAPIConnection api = new BoxAPIConnection("");
+
+        api.setConnectTimeout(instanceConnectTimeout);
+        api.setReadTimeout(instanceReadTimeout);
+
+        BoxAPIRequest req = new BoxAPIRequest(api, new URL("https://api.box.com/2.0/users/me"), "GET");
+
+        assertEquals(instanceConnectTimeout, req.getConnectTimeout());
+        assertEquals(instanceReadTimeout, req.getReadTimeout());
+    }
 }